Prevent infinite loop
Bug reported by Andrew Mayo, during his MCMC process the code would
hang. This is extra to the process dying error as reported by Lauren
Weiss.

One of the do-while loops did not have an iteration-based exit
condition, only a convergence criterion and with bad input was causing
an infinite loop.

This change adds a check with `MAX_ITER` and will exit the loop if the
maximum number of iterations is met.

I have a concern that the current value of 35 may not be enough for this
particular loop. It is possible that this value needs raising.

The code changes are in the subproject so check there for the actual
diff.